♻️ 9NKYP3 task: decompose transition verification
