+++ .agentplane/tasks/202605290551-9NKYP3/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 9NKYP3 task: Workflow transition service decomposition [202605290551-9NKYP3]
