+++ .agentplane/tasks/202605290551-9NKYP3/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T05:52:16.598Z
+
+## Task
+
+- Task: `202605290551-9NKYP3`
+- Title: Workflow transition service decomposition
+- Status: DOING
+- Branch: `task/202605290551-9NKYP3/workflow-transition-service-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290551-9NKYP3/README.md`
+
+## Verification
+
+- State: ok
+- Note: Workflow transition verification rendering/hash helpers extracted into workflow-transition-verification.ts; workflow-transition-service.ts reduced to 353 lines while preserving transition APIs.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T05:52:16.598Z
+- Branch: task/202605290551-9NKYP3/workflow-transition-service-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../task/shared/workflow-transition-service.ts     | 81 ++--------------------
+ .../shared/workflow-transition-verification.ts     | 74 ++++++++++++++++++++
+ 2 files changed, 80 insertions(+), 75 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
